材料を管理する処理をカスタムフックに切り出す
